Update patch set 1
Patch Set 1:

Alright, the man of rsync is not really meaningful so best is too try. I updated the playbook backup for sf-project.io and I'll compare duration of the next runs with the previous run's logs I have. I'm running the ansible-playbook with -vv so I have the time delta between a task start and stop.

Patch-set: 1