Use a new BUP_DIR by month
This change change the workflow to create a new BUP_DIR every
month. That means a full backup is performed every month and
incremental backup is performed during the month.

This also ease pruning of old backup by simply removing
the old BUP_DIR directories from the filesystem. This action
needs to be done manually.

Change-Id: I49e69082cc39c96ea9b43f66ac6375fb18d47d44