Update patch set 1
Patch Set 1:

Something is unclear to me. What is our source of info for the operations: taiga or sf-ops ? For instance how to resolve gerrit replication stuck is not in the taiga issues. So now as eng-on-duty I will have to check two "databases" to find issues/solutions. I'm not against putting this kind of info in sf-ops but then why not sf-docs then ? What are your thoughts Matthieu / Tristan / Javier / Jakub / Zoli

Patch-set: 1