Update patch set 1
Patch Set 1:

It was decided that our primary knowledge base should be the list of closed Taiga issues, so I would expect to have it there.

It does not hurt to have a file with "common stuff", though, but I would not merge this unless we have it in a Taiga issue too.

Patch-set: 1
Reviewer: Gerrit User 200 <200@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Workflow=0