zc-web-auth.py: SSO auth helper for zuul-client
Add python script to simplify SSO auth for zuul-client with upstream
(sfio) and downstream (sf.hosted) instances. The script will fetch
a valid access token after the user authenticates in the SSO via the
web browser.

This script is meant to be used until
https://review.opendev.org/c/zuul/zuul-client/+/831946 gets merged
upstream, as it provides the same functionality.

The script currently works with upstream (sfio, rdo) and support for
downstream is in progress.

Change-Id: Ibe54bb0303f2b943254aceabdfe304f524c13c2b